iomanager instrunction failed

+++ kernel/ia32/include/IOManager.h
@@ -10,7 +10,7 @@ public:
     ~IOManager()=default;
 
     uint8_t ReadPort(uint8_t portNumber);
-    void WritePort(uint8_t portNumber, uint8_t data);
+    void WritePort(uint8_t portNumber, uint16_t data);
 
 private:
